requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
dominate==2.6.0
Flask==2.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
google-api-core==2.7.1
google-auth==2.6.0
google-cloud-vision==2.7.1
googleapis-common-protos==1.55.0
greenlet==1.1.2
grpcio==1.44.0
grpcio-status==1.44.0
gunicorn==20.1.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
numpy==1.22.2
opencv-contrib-python==4.5.5.62
proto-plus==1.20.3
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
visitor==0.1.3
Werkzeug==2.0.3
google-cloud-language==2.4.1
google-cloud-bigquery==3.0.1
google-cloud-storage==2.2.1